backend/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Utilities
python-multipart>=0.0.6
Pillow>=10.0,<11.0
python-dateutil>=2.8,<3.0
pytz>=2024.1

# OpenAPI/Swagger
drf-spectacular>=0.27,<1.0

# Caching
django-redis>=5.4,<6.0

# Celery for async tasks
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0

# Redis
redis>=5.0,<6.0

# Validation
pydantic>=2.0,<3.0

# File storage
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Logging
python-json-logger>=2.0,<3.0

# HTTP
requests>=2.31,<3.0

# JWT
PyJWT>=2.8,<3.0

# SMS/Email
django-phonenumber-field>=7.2,<8.0
phonenumbers>=8.13,<9.0
dj-database-url>=2.1,<3.0
